+++ docs/CHANGELOG.md
@@ -1,3 +1,11 @@
+v0.1.8_RC2 (2025-01-22)
+-------------------
+
+- bug fix clusters/{area}/list.ini file was missing
+- bug fix for input/thermal/series/{area}/{cluster}/series.txt /data.txt and modulation.txt, wrong path 
+  at cluster creation
+
+
 v0.1.8_RC1 (2025-01-22)
 -------------------